Move final classes from test agent to engine core
